requirements.txt
python-bittrex